for the scanning of the item
if u need help-talk to me, the data that returns from the scanning is a string ('12212334') and then uses this to get the details in the DB, i also sent as props the superID so we can use the API u did so make sure when using this to pass it. also i added as props the function after getting the data about the item.